+++ DXGL-FRAMEWORK/Application/Source/SceneRhythm.h
@@ -17,7 +17,7 @@ public:
 	~SceneRhythm();
 
 	void Init() override;
-	void Update(double dt) override;
+	void Update(float dt) override;
 	void Render() override;
 	void Exit() override;
